requirements.txt
torch
torchvision
albumentations
pillow
tqdm
